R UI 222 improvements (#112)
* improvements

* improvements

* improve inputs - cleanup
