+++ src/components/editors/MeasureMultiSelectFieldPro/definition.ts
@@ -1,6 +1,6 @@
 import { Value } from '@embeddable.com/core';
 import { EmbeddedComponentMeta, Inputs, definePreview } from '@embeddable.com/react';
-import Component from './MeasureMultiSelectFieldPro';
+import Component from './index';
 import { inputs } from '../../component.inputs.constants';
 import { previewData } from '../../preview.data.constants';
 
@@ -12,11 +12,7 @@ const meta = {
   defaultHeight: 120,
   inputs: [
     inputs.dataset,
-    {
-      ...inputs.measureOptions,
-      label: 'Measure options (to show in dropdown)',
-      required: false,
-    },
+    inputs.measureOptions,
     inputs.title,
     inputs.description,
     inputs.tooltip,
@@ -68,8 +64,6 @@ const preview = definePreview(Component, {
 const props = (inputs: Inputs<typeof meta>) => {
   return {
     ...inputs,
-    measureOptions: inputs.measureOptions ?? [],
-    selectedMeasures: inputs.selectedMeasures ?? [],
   };
 };
